fix c2e return dtype

+++ py360convert/c2e.py
@@ -156,7 +156,7 @@ def c2e(
     coor_x_norm = (np.clip(coor_x, -0.5, 0.5) + 0.5) * face_w
     coor_y_norm = (np.clip(coor_y, -0.5, 0.5) + 0.5) * face_w
 
-    equirec = np.empty((h, w, cube_faces.shape[3]))
+    equirec = np.empty((h, w, cube_faces.shape[3]), dtype=cube_faces[0].dtype)
     for i in range(cube_faces.shape[3]):
         equirec[..., i] = sample_cubefaces(cube_faces[..., i], tp, coor_y_norm, coor_x_norm, order=order)
